Made changed to GUI
