+++ src/main/java/bobthebot/gui/DialogBox.java
@@ -15,6 +15,9 @@ public class DialogBox extends HBox {
     private static final int DISPLAY_PICTURE_WIDTH = 100;
     private static final int DISPLAY_PICTURE_HEIGHT = 100;
 
+    // @@author janelleljt-reused
+    // Reused from https://se-education.org/guides/tutorials/javaFx.html
+    // with minor modifications
     public DialogBox(Label l, ImageView iv) {
         text = l;
         displayPicture = iv;
@@ -27,6 +30,9 @@ public class DialogBox extends HBox {
         this.getChildren().addAll(text, displayPicture);
     }
 
+    // @@author janelleljt-reused
+    // Reused from https://se-education.org/guides/tutorials/javaFx.html
+    // with minor modifications
     /**
      * Flips the dialog box such that the ImageView is on the left and text on the right.
      */
@@ -37,10 +43,16 @@ public class DialogBox extends HBox {
         this.getChildren().setAll(tmp);
     }
 
+    // @@author janelleljt-reused
+    // Reused from https://se-education.org/guides/tutorials/javaFx.html
+    // with minor modifications
     public static DialogBox getUserDialog(Label l, ImageView iv) {
         return new DialogBox(l, iv);
     }
 
+    // @@author janelleljt-reused
+    // Reused from https://se-education.org/guides/tutorials/javaFx.html
+    // with minor modifications
     public static DialogBox getDukeDialog(Label l, ImageView iv) {
         var db = new DialogBox(l, iv);
         db.flip();
